Added plotting functionality. WARNING: Read Description
The plotting functionality currently has a bottleneck that needs to be addressed. To fix this, each UDP messsage should be stored in a log file, and then, chart.js will need to read that log file and retrieve the most recent value. This will significantly reduce the amount of power/memory consumption.